feat: extract Navbar into shared component
feat: extract Navbar into shared component

- Created src/components/Navbar.jsx with auth state and dropdown
- Removed navbar JSX and related state from Home.jsx
- Added Navbar to IoT dashboard and ChessTrack
